Adds health bar to entities
Implements a health bar UI element for entities.

The health bar is displayed above the entity and shows the current health as a percentage of the maximum health.
The health bar is only displayed for non-local and living entities.
The scale of the EntityBase prefab was increased for better visibility of the health bar.